delete constructor and rename
