+++ base/base.vcxproj
@@ -290,7 +290,7 @@
     <ClInclude Include="status.hpp" />
     <ClInclude Include="status_or.hpp" />
     <ClInclude Include="status_or_body.hpp" />
-    <ClInclude Include="type_traits.hpp" />
+    <ClInclude Include="not_constructible.hpp" />
     <ClInclude Include="unique_ptr_logging.hpp" />
     <ClInclude Include="unique_ptr_logging_body.hpp" />
     <ClInclude Include="version.generated.h" />
