+++ mathematica/mathematica_body.hpp
@@ -8,7 +8,7 @@
 #include <tuple>
 #include <vector>
 
-#include "base/type_traits.hpp"
+#include "base/not_constructible.hpp"
 
 namespace principia {
