+++ quantities/quantities.hpp
@@ -8,7 +8,7 @@
 #include <type_traits>
 
 #include "base/not_null.hpp"
-#include "base/type_traits.hpp"
+#include "base/not_constructible.hpp"
 #include "serialization/quantities.pb.h"
 
 namespace principia {
